third_party/WebKit/LayoutTests/webmidi/permission-expected.txt

Test if sysex permission request is handled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.requestMIDIAccess is defined.
PASS PermissionsHelper.setPermission is defined.
PASS promise is defined.
PASS promise.then is defined.
PASS sysex permission request is successfully rejected.
PASS promise is defined.
PASS promise.then is defined.
PASS sysex permission request is successfully accepted.
PASS access is defined.
PASS access.sysexEnabled is defined.
PASS access.sysexEnabled is true
PASS all permission request rests pass successfully.
PASS successfullyParsed is true

TEST COMPLETE
